Experimental re-design of BufferLine.
The new data structure is more flexible and more compact.
However, mapping column number to cell information is no longer O(1).
So we use a cursor as much as possible to reduce the need for that.

This is an initial check-in to have a backup and baseline.
It does NOT WORK except for a few simple things.